% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstfactor.R
\name{authelia_firstfactor}
\alias{authelia_firstfactor}
\title{Authelia authentication}
\usage{
authelia_firstfactor(
  self,
  private,
  username,
  password,
  domain = NULL,
  keepMeLoggedIn = TRUE
)
}
\arguments{
\item{self}{R6 self reference}

\item{private}{R6 private object}

\item{username}{The username to authenticate against}

\item{password}{The password to authenticate with}

\item{domain}{The domain to authenticate against. If left NULL, no redirect
header will be supplied}

\item{keepMeLoggedIn}{A boolean flag on whether to keep the user logged in.
This utilises the server's configuration for the duration of authentication}
}
\description{
This function passes a request to the authelia API to perform first-factor
authentication.
}
\keyword{internal}
