% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{authelia_init}
\alias{authelia_init}
\title{Initalize an authelia object}
\usage{
authelia_init(self, private, url)
}
\arguments{
\item{self}{R6 self reference}

\item{private}{R6 private object}

\item{url}{The URL of your authelia instance}
}
\description{
This is an internal function used to initialize a new authelia object
}
\keyword{internal}
