% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{authelia_verify}
\alias{authelia_verify}
\title{Authelia verification}
\usage{
authelia_verify(self, private, domain)
}
\arguments{
\item{self}{R6 self reference}

\item{private}{R6 private object}

\item{domain}{The URL of the domain we are verifying against}
}
\description{
This function passes a request to the authelia API to test whether the
requester is authenticated and verified to access a particular domain
}
\keyword{internal}
