% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autheliar.R
\docType{package}
\name{autheliar-package}
\alias{autheliar}
\alias{autheliar-package}
\title{autheliar: An R wrapper for the authelia API}
\description{
Autheliar helps to provide users with an intuitive interface to the
    Authelia API. It allows R users to implement SSO authentication with the goal
    focused on providing authentication to plumber APIs and shiny applications.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/liamgilbey/autheliar/issues}
}

}
\author{
\strong{Maintainer}: Liam Gilbey \email{liam.gilbey@gmail.com}

}
\keyword{internal}
